[JS][COM-51] Rename Particle class to PointMass + add include and src folder structure.
