+++ Physics/Physics.vcxproj
@@ -127,12 +127,12 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="core.h" />
-    <ClInclude Include="particle.h" />
-    <ClInclude Include="precision.h" />
+    <ClInclude Include="include\core.h" />
+    <ClInclude Include="include\pointmass.h" />
+    <ClInclude Include="include\precision.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="particle.cpp" />
+    <ClCompile Include="src\pointmass.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
