+++ Physics/Physics.vcxproj.filters
@@ -15,18 +15,18 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="core.h">
+    <ClInclude Include="include\core.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="precision.h">
+    <ClInclude Include="include\precision.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="particle.h">
+    <ClInclude Include="include\pointmass.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="particle.cpp">
+    <ClCompile Include="src\pointmass.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
